requirements.txt
# Core dependencies
boto3>=1.26.0
botocore>=1.29.0
click>=8.1.0
pyyaml>=6.0
jinja2>=3.1.0
tabulate>=0.9.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Bioinformatics
biopython>=1.81
pysam>=0.21.0

# Reporting
reportlab>=4.0.0
pypdf>=3.8.0
markdown>=3.4.0
weasyprint>=59.0

# Web framework (for dashboard)
streamlit>=1.22.0
fastapi>=0.95.0
uvicorn>=0.22.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
moto>=4.1.0  # AWS service mocking

# Development tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.2.0
pre-commit>=3.3.0

# Monitoring
prometheus-client>=0.16.0
datadog>=0.45.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.10.0

# AWS specific
aws-lambda-powertools>=2.15.0
aws-cdk-lib>=2.80.0

# Utilities
python-dotenv>=1.0.0
requests>=2.30.0
tenacity>=8.2.0  # Retry logic
tqdm>=4.65.0  # Progress bars

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
mkdocs>=1.4.0
mkdocs-material>=9.1.0